Corrigindo tipos das primary keys

+++ src/main/java/com/cotemig/repositories/CondoRepository.java
@@ -3,6 +3,6 @@ package com.cotemig.repositories;
 import com.cotemig.models.Condo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CondoRepository extends JpaRepository<Condo, Long> {
+public interface CondoRepository extends JpaRepository<Condo, Integer> {
 
 }
