+++ src/main/java/com/cotemig/repositories/ResidentRepository.java
@@ -3,6 +3,6 @@ package com.cotemig.repositories;
 import com.cotemig.models.Resident;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResidentRepository  extends JpaRepository<Resident, Long> {
+public interface ResidentRepository  extends JpaRepository<Resident, Integer> {
 
 }
